feat: Reorganize repo structure and update README.md
- Move CommandFactory.py, CommandHandler.py, hail_methods.py, utils.py, file_utility.py, and stats.py to src folder
- Update import statements in main.py and other affected files to match new directory structure
- Modify README.md to reflect changes in repo organization and provide information on additional files
- Relocate stats.py to src folder as it is part of the project's source code
