+++ main.py
@@ -2,11 +2,9 @@ from pathlib import Path
 import argparse
 import hail as hl
 from pyspark import SparkConf, SparkContext
-from hail_methods import import_and_annotate_vcf, merge_matrix_tables, reduce_to_2d_table, create_frequency_bins
 
-
-from CommandFactory import CommandFactory
-from CommandHandler import CommandHandler
+from src.CommandFactory import CommandFactory
+from src.CommandHandler import CommandHandler
 
 hail_home = Path(hl.__file__).parent.__str__()
 if __name__ == '__main__':
